chore(*) rename 'core' to 'runloop'
With the upcoming plugins SDK and `kcore`, `kong` globals, this is an
attempt at reducing the confusion around various components.

While historically, the `kong/core/` directory was implemented as a
plugin, it has deviated quite a lot from this original intent over time.

Modules belonging to the `runloop` namespaces are top-level logic
executed within the NGINX phases (i.e. not plugins).

As of today, the runloop is still unfortunately split between
`kong/init.lua` and `kong/runloop/`, but this is a task for another day.
