+++ spec/01-unit/011-balancer_spec.lua
@@ -17,7 +17,7 @@ describe("Balancer", function()
   setup(function()
     stub(ngx, "log")
 
-    balancer = require "kong.core.balancer"
+    balancer = require "kong.runloop.balancer"
     singletons = require "kong.singletons"
     singletons.worker_events = require "resty.worker.events"
     singletons.dao = {}
